refactor: #115 - postpage Hooks폴더의 index.ts 파일 추가
