+++ src/pages/PostPage/Hooks/index.ts
@@ -1,0 +1,4 @@
+export { default as useCommentNotification } from './useCommentNotification';
+export { default as useCreateComment } from './useCreateComment';
+export { default as useDeleteComment } from './useDeleteComment';
+export { default as useVotingResult } from './useVotingResult';
