+++ src/pages/PostPage/components/Turnout.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Comment } from '@type';
 import { ANGOLA_STYLES } from '@styles/commonStyles';
-import useVotingResult from '../Hooks/useVotingResult';
+import { useVotingResult } from '../Hooks';
 import { VOTED_VALUES } from '../constants';
 
 interface TurnoutProps {
